refactor: redefinition `ShopReview`
