+++ src/models/auth/Review.ts
@@ -9,4 +9,20 @@ export type Review = {
   updatedAt: number;
 };
 
+export type ShopReview = {
+  id: number;
+  shop: {
+    name: string;
+    category: string;
+  };
+  profile: {
+    nickname: string;
+    profileImageSrc: string;
+  };
+  content: string;
+  imgSrc: string[];
+  updatedAt: number;
+};
+
 export type ReviewList = Review[];
+export type ShopReviewList = ShopReview[];
